+++ repositories/course_repository.py
@@ -1,0 +1,4 @@
+from repositories.repository import Repository
+
+class CourseRepository(Repository):
+    pass