+++ repositories/database_course_repository.py
@@ -1,0 +1,14 @@
+class DatabaseCourseRepository:
+
+    def save(self, course):
+        # Future database implementation
+        pass
+
+    def find_by_id(self, course_id):
+        pass
+
+    def find_all(self):
+        pass
+
+    def delete(self, course_id):
+        pass